Finish desired styling for project
The last commit was a bit vague in terms of what was added. I fixed a
CSS-related issue where, under certain zoom magnitudes, the content
of the article area would overflow the article container. I have added
medisa rules that make the sidebar disappear when the max-width
of a screen is under a certain value. I plan to re-add the sidebar
through use of JavaScript. The color scheme has been greatly improved
as well.